Potion Throwing
Improved the backend calculations for brewing, and implemented the ability to throw/consume whatever potion you're holding. It doesn't apply the effect right now, but that's what I'll be implementing next. I also added a blank panel to kind of get an idea for where the notebook could be when you open it up.