Potion Drinking + Throwing
Allows for Potion throwing and drinking, and they each apply effects based on what was used int he potion.

There is also the chance that you may encounter a bug where the ActiveGenericPotion scene reportedly corrupts and cannot be opened in the editor. It still works when running the game though, and you can ask me on how to fix the issue. Just try reloading the project scene until then.